requirements.txt
Django==2.2.6
djangorestframework==3.12.2
mysql-connector-python==8.0.22
protobuf==3.14.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1